[*] Refactoring types

+++ src/wasm/constants/mod.rs
@@ -1,9 +1,9 @@
-pub const MAX_H: i32 = 360;
-pub const MAX_SVL: i32 = 100;
-pub const MIN_HSV: i32 = 0;
+pub const MAX_H: f32 = 360.0;
+pub const MAX_SVL: f32 = 100.0;
+pub const MIN_HSV: f32 = 0.0;
 
-pub const MIN_RGB: i32 = 0;
-pub const MAX_RGB: i32 = 255;
+pub const MIN_RGB: f32 = 0.0;
+pub const MAX_RGB: f32 = 255.0;
 
-pub const MIN_CMYK: i32 = 0;
-pub const MAX_CMYK: i32 = 100;
+pub const MIN_CMYK: f32 = 0.0;
+pub const MAX_CMYK: f32 = 100.0;
